MD3UP: aim correctly when doing mesh scans
The MD3UP's 'raster scan' command is relative the last save
centered position. Make sure we have mesh center saved as
last centered position before issuing the raster scan command.

Also, give correct width and height of the scan area to MD3Up.
The are shoud span the center of the outer cells, not the edges.