wait longer for detector cover to open
It takes more then 1 second to open the cover. Wait 5 seconds for
now.

This needs to be changed so that we actually check that cover is open,
instead of waiting some guesstimated ammount of time.

There is an tango attribute B312A_E06_DIA_DETC02_ENAS can be used
to check if cover is open.