skip wait_ready() for Jungfrau detector
A hack to deal with missmatch between how Eiger and Jungfrau
handles states.

Eiger goes into 'ready' state when it's armed and ready for data
acquisition.

Jungfrau goes into 'busy' state when it's armed and stays in 'busy'
state during data acquisition. Jungfrau goes into 'ready' when it's
is ready to be armed, that is whent it is idle.