YAML: use 'global' reference to centering HWOBJ in MD3
As now it's not possible to load hardware object twice from same
configuration file, we have a circular dependency between MD3 and
centering hardware object.

The MD3 uses a reference to centering hardware object to invoke
it's methods. The centering hardware object loads reference to
MD3 motors.

Deal with this by using a 'global' reference to centering HWOBJ
in MD3 code. This way centering can be loaded _after_ MD3 is
loaded.