set 'create_point' flag for SSX tasks
Adds a new collection parameter 'create_flag'. Set it to true
for SSX tasks.

This implements same behaviour for SSX tasks as for the built-in
'Data Collection (Limited OSC)' task. I.e. a 2D point is created
where user clicked, and the collection is performed at that point.

Note, that this requires support in the front-end as well.