check if Beamstop is in 'BEAM' position
When doing data collections, after MD3 was put into 'DataCollection',
check that Beamstop was moved in.

Check that MD3 reports that Beamstop is at 'BEAM' position. If
not, abort data collection, in an attempt to not shoot the detector
with direct X-ray beam.

This check is an attempt to address some kind of issues with MD3Up,
where it sometimes does not move in Beamstop.