ISARA: deal with tango attribute connection errors
Check if we actually manage to access the attributes of ISARA's
tango device.

If there is an error accessing attributes during initialization,
put the ISARA hardware object into fault mode and abort
initialization.

We must make sure that ISARA hardware object is successfully
created and is not None. MXCuBE invokes methods on the sample
changer HWOBJ during login. If sample changer HWOBJ is None,
the login will fail.