Jungfrau SSX: set OmegaIncrement to non-zero value
When doing our SSX Injector experiments, we don't rotate the omega
axis, thus omega increments IS 0. However Jungfrau consider 0 omega
increment an invalid setting, and will refuse to arm. Set omega
increment to a 'small' angle to placate it.

This have probably been changed in later versions of tango device
or firmware of the Jungfrau detector.