MICROMAXCOLLECT: Now properly handles GreenletExit
There was an issue, where due to moving out of detector distance
value limits, the data collection stopped, but the task itself never
terminated in MXCuBE.

This was caused, because the `MICROMAXCollect.ready_event` was never
set, and the call to `ready_event.wait()` in the `collect` method never
terminated as a result.
Normally setting the event is done when the exception inheriting from
`Exception` is caught, but not a GreenletExit.
GreenletExit inherits from BaseException instead of Exception, thus no
exception handler caught it.