Stops collection once when detector moves out of limits
The previous implementation if `MICROMAXCollect.move_detector` first
checked if there was an attempt to move the detector out of values
specified by limits. In that case `stop_collect` was called, and then
the function was further executed.
Then `stop_collect` was executed for the second time, because detector did not
move to the specified position.

Now `stop_collect` will be called at most once.