MICROMAXCollect: fix taking crystal snapshots
Currently this feature is not working. Setting number of snapshots to be
taken was not respected. Decision to take snapshots or not was made
based on `take_snapshots` data collection parameter.

This fix works by using `collect.number_of_snapshots` parameter instead,
which is the one set by the web UI.

This also required another small fix.

The camera hardware object was accessed by a wrong property.
The reference `self.camera` was assigned only in some method of the
parent class. Switched to use HWR instead.

WIP: make attribute in __init__()

TODO: squash!