Migrate BioMAX Isara from InSoak to PositionName
The Tango device server for Isara sample changers
now has a `PositionName` attribute for Isara1 as well.
`PositionName` delivers a `SOAK` value when the gripper is soaking.
So `PositionName` can be used instead of the `InSoak` attribute.
This is better for consistency since `InSoak` is a made up attribute,
while `PositionName` at least is a real value on Isara2.