MD3: remove 'focus' and 'zoom' from centring motors list
The 'focus' and 'zoom' should not be included into the
'centring motors' list. The 'centring motors' are used to
calculate shape's motor positions dict.

For example, without this change, focus and zoom position are
included into 2D points position. That is not correct, as we
don't want to change current zoom and focus when moving to a
2D point.