SSX: No longer overwrites fields values
update_dependent_fields method is called when form state is updated on
the frontend.
It's goal is to return new values for some form fields.
Then these new values are being set as new values of the input fields on
the frontend.

It should be used when value of some field stored in the form is
**dependent** on values of other field (in that case such input would be
usually hidden).

Currently it is called with all form fields' values as an input and
returns the same, unmodified values.
It makes it so that each input field for these values gets overwritten
with the value we get.
It leads to a situation when input with `0.0` value gets overwritten
with `0` on the frontend.
It was caused by these actions:
Value `0.0` is passed into `Number` constructor on the fronted, which
yiels a `0` value. Then this value is sent to `update_dependent_fields`
method, which returns the `0` value again. This `0` values is set as the
input fields' value via DOM manipulation.

This change makes it so that no field is being overwritten, as we don't
have 'dependent' fields in the forms for SSX tasks.