Adds default value to `space_group` parameter
It fixes some validation issue on the mxcube-web.
Without default value, the field is not validated before some value is
assigned.
It is not present in `formData` which stores all the form
values.
This results in space group value not being set initially and user being
able to run the task without specifying value of this field.