Refactor header_appendix code into parent class
Both `collect` objects implements a way to create a header_appendix
json-string.
The logic is pretty similar overall, with slight differences.

`BioMAX` has the following keys that are not present in `MicroMAX collect_dict`:
- `cell_counting`
- `start_corner`
- `mesh_range`
- `scan_pattern`
- `scan_orienttation`

And `MicroMAX` has the following keys that are not present in `BioMAX collect_dict`:
- `ssx_mode`
- `target_beam_size_factor`

I decided to create a `_create_header_appendix` method which creates the
dictionary containing the shared portion of the `header_appendix` dict.
Then a base class may override it, to extend it with new entries.