YAML: MD3: load CentringMath reference lazily
Due to way hardware object are now loaded it's not possble to
'attache' same hardware objects at multiple points. Thus it's
not possible to have 'centring' hardware object to be attached
to both 'Beamline' and 'MAXIVMD3' hardware objects.

Deal with this by lazily creating reference to 'centring' on first
use.