ISARA: handle more error cases from tango device
Make ISARA and ISARAMaint hardware object handle more error
types returned by the isara2 tango device.

Now when setting up polling of isara2 attributes:

 * check if channel object is connected
 * try to read the attribute
 * catch DevFailed exceptions

If any of the tests fail, we put sample changer into the fault
mode.

This make it possible to use MXCuBE when the sample changer robot
is unavailable, for example turned off. This will be usable at for
example MicroMAX, as there multiple experiments that don't use the
robot.