Fix and refactor file names for diffraction images
This moves the code handling diffraction images to `MAXIV.DataCollect`.

Change is intended to be backwards-compatible for MicroMAX,
only refactoring.

Change is intended to fix BioMAX, where it was broken before.
The frame number was not taken into account.

This also improves slightly the handling of SSH requests for this.