MicroMAX flux: use global reference to energy HWO
Use global reference for accessing the energy hardware object.
It's no longer possible to attach energy to flux object. It
needs to be attached to the `beamline` object now.

Note, it's not possible to set-up the old `self.self.energy_hwobj`
in the `init()` method, as flux object is loaded _before_
energy object. Thus energy does not yet exist when we are
running `init()`.